- Se agregaron modelos. - Se modificaron las migraciones, para produccion.

+++ database/migrations/2025_05_17_010806_create_planes_suscripciones_table.php
@@ -27,6 +27,9 @@ return new class extends Migration {
      */
     public function down(): void
     {
+        if (app()->environment('production')) {
+            throw new Exception('The "down" method is disabled in production.');
+        }
         Schema::dropIfExists('planes_suscripciones');
     }
 };
