+++ database/migrations/2025_05_17_010848_create_historial_pagos_suscripciones_table.php
@@ -26,6 +26,9 @@ return new class extends Migration {
      */
     public function down(): void
     {
+        if (app()->environment('production')) {
+            throw new Exception('The "down" method is disabled in production.');
+        }
         Schema::dropIfExists('historial_pagos_suscripciones');
     }
 };
