+++ database/migrations/2025_05_17_010945_create_usuarios_table.php
@@ -32,6 +32,9 @@ return new class extends Migration {
      */
     public function down(): void
     {
+        if (app()->environment('production')) {
+            throw new Exception('The "down" method is disabled in production.');
+        }
         Schema::dropIfExists('usuarios');
     }
 };
