+++ database/migrations/2025_05_17_011040_create_items_menus_table.php
@@ -29,6 +29,9 @@ return new class extends Migration {
      */
     public function down(): void
     {
+        if (app()->environment('production')) {
+            throw new Exception('The "down" method is disabled in production.');
+        }
         Schema::dropIfExists('items_menus');
     }
 };
