+++ database/migrations/2025_05_17_011312_create_facturas_table.php
@@ -31,6 +31,9 @@ return new class extends Migration {
      */
     public function down(): void
     {
+        if (app()->environment('production')) {
+            throw new Exception('The "down" method is disabled in production.');
+        }
         Schema::dropIfExists('facturas');
     }
 };
